requirements.txt
flask
opencv-python
pillow
numpy
ultralytics
